spli cli
