+++ packages/cli/src/cli.ts
@@ -1,0 +1,248 @@
+import {
+    RequestError,
+    CORE_VERSION,
+    isRequestError,
+    TOOL_ID,
+    TOOL_NAME,
+    GITHUB_REPO,
+    SERVER_PORT,
+    RUNTIME_ERROR_CODE,
+    UNHANDLED_ERROR_CODE,
+} from "genaiscript-core"
+import { NodeHost } from "./nodehost"
+import { program } from "commander"
+import { error, isQuiet, setConsoleColors, setQuiet } from "./log"
+import { startServer } from "./server"
+import { satisfies as semverSatisfies } from "semver"
+import { NODE_MIN_VERSION } from "./version"
+import { runScript } from "./run"
+import { buildProject } from "./build"
+import { batchScript } from "./batch"
+import {
+    codeOutline,
+    retreivalClear,
+    retreivalIndex,
+    retreivalSearch,
+    retreivalTokens,
+} from "./retreival"
+import { helpAll } from "./help"
+import { jsonl2json, parseDOCX, parseFence, parsePDF } from "./parse"
+
+async function listScripts() {
+    const prj = await buildProject()
+    console.log("id, title, group, filename, system")
+    prj.templates.forEach((t) =>
+        console.log(
+            `${t.id}, ${t.title}, ${t.group || ""}, ${t.filename || "builtin"}, ${
+                t.isSystem ? "system" : "user"
+            }`
+        )
+    )
+}
+
+export async function cli() {
+    process.on("uncaughtException", (err) => {
+        error(isQuiet ? err : err.message)
+        if (isRequestError(err)) {
+            const exitCode = (err as RequestError).status
+            process.exit(exitCode)
+        } else process.exit(UNHANDLED_ERROR_CODE)
+    })
+
+    if (!semverSatisfies(process.version, NODE_MIN_VERSION)) {
+        console.error(
+            `node.js runtime incompatible, expected ${NODE_MIN_VERSION} got ${process.version}`
+        )
+        process.exit(RUNTIME_ERROR_CODE)
+    }
+
+    program.hook("preAction", (cmd) => {
+        NodeHost.install(cmd.opts().env)
+    })
+
+    program
+        .name(TOOL_ID)
+        .version(CORE_VERSION)
+        .description(`CLI for ${TOOL_NAME} ${GITHUB_REPO}`)
+        .showHelpAfterError(true)
+        .option("--env <path>", "path to .env file, default is './.env'")
+        .option("--no-colors", "disable color output")
+        .option("-q, --quiet", "disable verbose output")
+    program.on("option:no-colors", () => setConsoleColors(false))
+    program.on("option:quiet", () => setQuiet(true))
+
+    program
+        .command("run")
+        .description("Runs a GenAIScript against files or stdin.")
+        .arguments("<script> [files...]")
+        .option("-ef, --excluded-files <string...>", "excluded files")
+        .option(
+            "-o, --out <string>",
+            "output folder. Extra markdown fields for output and trace will also be generated"
+        )
+        .option("-rmo, --remove-out", "remove output folder if it exists")
+        .option("-ot, --out-trace <string>", "output file for trace")
+        .option(
+            "-od, --out-data <string>",
+            "output file for data (.jsonl/ndjson will be aggregated). JSON schema information and validation will be included if available."
+        )
+        .option(
+            "-oa, --out-annotations <string>",
+            "output file for annotations (.csv will be rendered as csv, .jsonl/ndjson will be aggregated)"
+        )
+        .option("-ocl, --out-changelog <string>", "output file for changelogs")
+        .option("-j, --json", "emit full JSON response to output")
+        .option("-y, --yaml", "emit full YAML response to output")
+        .option(
+            "-p, --prompt",
+            "dry run, don't execute LLM and return expanded prompt"
+        )
+        .option(`-fe, --fail-on-errors`, `fails on detected annotation error`)
+        .option("-r, --retry <number>", "number of retries", "8")
+        .option(
+            "-rd, --retry-delay <number>",
+            "minimum delay between retries",
+            "15000"
+        )
+        .option(
+            "-md, --max-delay <number>",
+            "maximum delay between retries",
+            "180000"
+        )
+        .option("-l, --label <string>", "label for the run")
+        .option("-t, --temperature <number>", "temperature for the run")
+        .option("-tp, --top-p <number>", "top-p for the run")
+        .option("-m, --model <string>", "model for the run")
+        .option("-mt, --max-tokens <number>", "maximum tokens for the run")
+        .option("-se, --seed <number>", "seed for the run")
+        .option("--no-cache", "disable LLM result cache")
+        .option("--cs, --csv-separator <string>", "csv separator", "\t")
+        .option("-ae, --apply-edits", "apply file edits")
+        .option(
+            "--vars <namevalue...>",
+            "variables, as name=value, stored in env.vars"
+        )
+        .action(runScript)
+
+    program
+        .command("batch")
+        .description("Run a tool on a batch of specs")
+        .arguments("<script> [files...]")
+        .option("-ef, --excluded-files <string...>", "excluded files")
+        .option(
+            "-o, --out <folder>",
+            "output folder. Extra markdown fields for output and trace will also be generated"
+        )
+        .option("-rmo, --remove-out", "remove output folder if it exists")
+        .option("-os, --out-summary <file>", "append output summary in file")
+        .option("-r, --retry <number>", "number of retries", "8")
+        .option(
+            "-rd, --retry-delay <number>",
+            "minimum delay between retries",
+            "15000"
+        )
+        .option(
+            "-md, --max-delay <number>",
+            "maximum delay between retries",
+            "180000"
+        )
+        .option("-l, --label <string>", "label for the run")
+        .option("-t, --temperature <number>", "temperature for the run")
+        .option("-tp, --top-p <number>", "top-p for the run")
+        .option("-m, --model <string>", "model for the run")
+        .option("-mt, --max-tokens <number>", "maximum tokens for the run")
+        .option("-se, --seed <number>", "seed for the run")
+        .option("--no-cache", "disable LLM result cache")
+        .option("-ae, --apply-edits", "apply file edits")
+        .option(
+            "--vars <string...>",
+            "variables, as name=value, stored in env.vars"
+        )
+        .action(batchScript)
+
+    program
+        .command("scripts")
+        .description("List all available scripts in workspace")
+        .action(listScripts)
+
+    const retreival = program.command("retreival").description("RAG support")
+    retreival
+        .command("index")
+        .description("Index a set of documents")
+        .argument("<file...>", "Files to index")
+        .option("-ef, --excluded-files <string...>", "excluded files")
+        .option("-n, --name <string>", "index name")
+        //        .option("-s, --summary", "use LLM-generated summaries")
+        .option("-cs, --chunk-size <number>", "chunk size")
+        .option("-co, --chunk-overlap <number>", "chunk overlap")
+        .option("-m, --model <string>", "model for embeddings (default gpt-4)")
+        .option(
+            "-sls, --split-long-sentences",
+            "split long sentences (default true)"
+        )
+        .action(retreivalIndex)
+    retreival
+        .command("search")
+        .description("Search index")
+        .arguments("<query> [files...]")
+        .option("-ef, --excluded-files <string...>", "excluded files")
+        .option("-tk, --top-k <number>", "maximum number of embeddings")
+        .option("-n, --name <string>", "index name")
+        //        .option("-s, --summary", "use LLM-generated summaries")
+        .action(retreivalSearch)
+    retreival
+        .command("clear")
+        .description("Clear index to force re-indexing")
+        .option("-n, --name <string>", "index name")
+        //        .option("-s, --summary", "use LLM-generated summaries")
+        .action(retreivalClear)
+
+    retreival
+        .command("outline")
+        .description("Generates a compact code repository outline")
+        .arguments("<files...>")
+        .option("-ef, --excluded-files <string...>", "excluded files")
+        .action(codeOutline)
+
+    program
+        .command("serve")
+        .description("Start a GenAIScript local server")
+        .option(
+            "-p, --port <number>",
+            `Specify the port number, default: ${SERVER_PORT}`
+        )
+        .action(startServer)
+
+    const parser = program.command("parse").description("Parse various outputs")
+    parser
+        .command("fence <language>")
+        .description("Extracts a code fenced regions of the given type")
+        .action(parseFence)
+
+    parser
+        .command("pdf <file>")
+        .description("Parse a PDF into text")
+        .action(parsePDF)
+
+    parser
+        .command("docx <file>")
+        .description("Parse a DOCX into texts")
+        .action(parseDOCX)
+    parser
+        .command("tokens")
+        .description("Count tokens in a set of files")
+        .arguments("<files...>")
+        .option("-ef, --excluded-files <string...>", "excluded files")
+        .action(retreivalTokens)
+    parser
+        .command("jsonl2json", "Converts JSONL files to a JSON file")
+        .argument("<file...>", "input JSONL files")
+        .action(jsonl2json)
+
+    program
+        .command("help-all", { hidden: true })
+        .description("Show help for all commands")
+        .action(helpAll)
+
+    program.parse()
+}
